Updating the UX to show status
