+++ pkg/scanner/scanner.go
@@ -65,7 +65,26 @@ func (s *Scanner) initProgress() {
 	s.Progress = map[string]progresstypes.Progress{}
 
 	for _, enabledCheck := range s.EnabledChecks {
+		// not all checks will use job segmentation for status
 		if enabledCheck == "unstable-github-ref" {
+			progress := progresstypes.Progress{}
+			for _, jobName := range s.JobNames {
+				progress.Set(jobName, false, false)
+			}
+			s.Progress[enabledCheck] = progress
+		} else if enabledCheck == "unfork-action" {
+			progress := progresstypes.Progress{}
+			for _, jobName := range s.JobNames {
+				progress.Set(jobName, false, false)
+			}
+			s.Progress[enabledCheck] = progress
+		} else if enabledCheck == "unstable-docker-tag" {
+			progress := progresstypes.Progress{}
+			for _, jobName := range s.JobNames {
+				progress.Set(jobName, false, false)
+			}
+			s.Progress[enabledCheck] = progress
+		} else if enabledCheck == "outdated-action" {
 			progress := progresstypes.Progress{}
 			for _, jobName := range s.JobNames {
 				progress.Set(jobName, false, false)
@@ -105,7 +124,8 @@ func (s *Scanner) ScanWorkflow() error {
 				s.RemediatedContent = updated
 			}
 		} else if check == "unstable-docker-tag" {
-			issues, err := unstabledockertag.DetectIssues(parsedWorkflow)
+			progress := s.Progress[check]
+			issues, err := unstabledockertag.DetectIssues(parsedWorkflow, progress.Set)
 			if err != nil {
 				return errors.Wrap(err, "failed to run unstable unstable-docker-tag check")
 			}
@@ -120,7 +140,8 @@ func (s *Scanner) ScanWorkflow() error {
 				s.RemediatedContent = updated
 			}
 		} else if check == "outdated-action" {
-			issues, err := outdatedaction.DetectIssues(parsedWorkflow)
+			progress := s.Progress[check]
+			issues, err := outdatedaction.DetectIssues(parsedWorkflow, progress.Set)
 			if err != nil {
 				return errors.Wrap(err, "failed to run unstable outdated-action check")
 			}
@@ -135,7 +156,8 @@ func (s *Scanner) ScanWorkflow() error {
 				s.RemediatedContent = updated
 			}
 		} else if check == "unfork-action" {
-			issues, err := unforkaction.DetectIssues(parsedWorkflow)
+			progress := s.Progress[check]
+			issues, err := unforkaction.DetectIssues(parsedWorkflow, progress.Set)
 			if err != nil {
 				return errors.Wrap(err, "failed to run unstable unfork-action check")
 			}
